Fix PR notes
